Conditional redis flush on startup

+++ pastamaker/worker.py
@@ -415,7 +415,8 @@ def error_handler(job, *exc_info):
 
 def main():
     utils.setup_logging()
-    utils.get_redis().flushall()
+    if config.FLUSH_REDIS_ON_STARTUP:
+        utils.get_redis().flushall()
     with rq.Connection(utils.get_redis()):
         worker = rq.worker.HerokuWorker([rq.Queue('default')],
                                         exception_handlers=[error_handler])
